Add more tests for constraint parser
