+++ lib/patterns_repository/models/throwable.dart
@@ -7,6 +7,8 @@ abstract class Throwable
     with Comparable<Throwable>, Compare<Throwable>, EquatableMixin {
   const Throwable();
 
+  // TODO: handle cases like height > 2, ...
+
   Fraction? get height;
   int? get passingIndex;
